% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_app_server.R
\name{app_server}
\alias{app_server}
\title{The application Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}
}
\value{
A Shiny Server object
}
\description{
This function sets up the data loading and integrates the modules for
the IsoRanker Shiny application. It is called internally by run_app.
}
